Moved the code from DcExitHandler.cs and SetSessionScene.cs into a combined file, DcUnityEvents.cs. Updated scenes to use new script instead of old ones. Moved functionality from FixNullStudent.cs into DcDataLogging.cs. Deleted some unnecessary files in the data collection folder.
