+++ Assets/Scenes/Egg Sort.unity
@@ -19110,8 +19110,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1119199311}
   - component: {fileID: 1119199310}
-  - component: {fileID: 1119199312}
-  - component: {fileID: 1119199313}
   m_Layer: 0
   m_Name: Data Collection
   m_TagString: Untagged
@@ -19146,30 +19144,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 11
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1119199312
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1119199309}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f47d311e0e75a77428bb41ba3a366b91, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1119199313
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1119199309}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0de5fcb0420515c40b6f72cbf7d62e28, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1122588704
 GameObject:
   m_ObjectHideFlags: 0
