+++ Assets/Scenes/Module_01.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 0.098
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.024569137, g: 0.035562772, b: 0.043603145, a: 0.098}
+  m_IndirectSpecularColor: {r: 0.024097772, g: 0.035355892, b: 0.044294912, a: 0.098}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -22105,8 +22105,6 @@ GameObject:
   - component: {fileID: 1527160400}
   - component: {fileID: 1527160399}
   - component: {fileID: 1527160401}
-  - component: {fileID: 1527160402}
-  - component: {fileID: 1527160403}
   m_Layer: 0
   m_Name: Data Collection
   m_TagString: Data
@@ -22151,30 +22149,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: b52b591b72d54744a4cfe6c6f69ea035, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1527160402
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1527160398}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f47d311e0e75a77428bb41ba3a366b91, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1527160403
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1527160398}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0de5fcb0420515c40b6f72cbf7d62e28, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &1532345386
