+++ Assets/Scenes/Tutorial.unity
@@ -39392,8 +39392,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1824096250}
   - component: {fileID: 1824096249}
-  - component: {fileID: 1824096251}
-  - component: {fileID: 1824096252}
   - component: {fileID: 1824096253}
   m_Layer: 0
   m_Name: Data Collection
@@ -39429,30 +39427,6 @@ Transform:
   m_Father: {fileID: 997847364}
   m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1824096251
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1824096248}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f47d311e0e75a77428bb41ba3a366b91, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1824096252
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1824096248}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0de5fcb0420515c40b6f72cbf7d62e28, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &1824096253
 MonoBehaviour:
   m_ObjectHideFlags: 0
