--- Assets/Scripts/DataCollection/Converters.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 21f35e41698ce1247817fd52a9eb7b92
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
