--- Assets/Scripts/DataCollection/Converters/DateTimeConverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7f5237b102e657c44a39e14fabcc5bfa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
