--- Assets/Scripts/DataCollection/Converters/DateTimeNullableConverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7085bff277115db4d840223f6b9042cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
