--- Assets/Scripts/DataCollection/DCTemplate.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b3d1924b10534655b798e195bcf0658b
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
