--- Assets/Scripts/DataCollection/DcExitHandler.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-namespace DataCollection
-{
-    public class DcExitHandler : MonoBehaviour
-    {
-        private void OnApplicationQuit()
-        {
-            DcDataLogging.EndSession();
-        }
-
-        public void EndSession()
-        {
-            DcDataLogging.EndSession();
-        }
-    }
-}
