--- Assets/Scripts/DataCollection/DcTemplate2.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1b3b300cb6244ce18869878fb4d2c979
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
