--- Assets/Scripts/DataCollection/FixNullStudent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f47d311e0e75a77428bb41ba3a366b91
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
