--- Assets/Scripts/DataCollection/SetSessionScene.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-using DataCollection;
-
-
-namespace DataCollection
-{
-    public class SetSessionScene : MonoBehaviour
-    {
-        // Start is called before the first frame update
-        void Start()
-        {
-            DcDataLogging.Session.SessionScene = SceneManager.GetActiveScene().name;
-        }
-    }
-}
-
