--- Assets/Scripts/DataCollection/SetSessionScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0de5fcb0420515c40b6f72cbf7d62e28
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
