--- Assets/Scripts/DataCollection/TestOutput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b9b2304bb4872e645bd143dab4c8da2a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
